test(payments): split cancellation emails out of the dunning suite
Move confirmation, copy, retry-scan, and shared-ledger cases into
cancellationEmails.test.ts so dunningEmails.test.ts stays on the
dunning + winback lifecycle.
